refactor(doctor): extract history and workspace checks
